+++ test/db/init.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS value_table (
     id INT AUTO_INCREMENT PRIMARY KEY,
-    value INT NOT NULL,
+    value FLOAT NOT NULL,
     timestamp DATETIME DEFAULT CURRENT_TIMESTAMP
 );
 
-INSERT INTO value_table (value) VALUES (50); -- 初期値
+INSERT INTO value_table (value) VALUES (50.0); -- 初期値
